Adjustments and Build
A lot of the changes I was making for the rectangular stages were implemented in a different way and the UI was bigger than I expected, so I left the square stages and instead increased the maximum size and set the camera to zoom out to match. I also made an adjustment to level gen that prevents too many obstacles from being made (i.e. less than half the map will be blocked off).
I also included the final build.